pyproject.toml
[project]
name = "theme-explorer"
version = "0.0.1"
description = "Exploratory tool for user feedback"
readme = "README.md"
authors = [
  { name = "Jyldyz Djumalieva" },
]
license = {text = "MIT License"}
classifiers = [
    "Programming Language :: Python :: 3.12",
]
requires-python = "==3.12.*"
dependencies = [
    "ipywidgets>=8.1.5",
    "ipykernel>=6.29.5",
    "nbstripout>=0.8.1",
    "pre-commit-hooks>=5.0.0",
    "pre-commit>=4.0.1",
    "ruff>=0.8.4",
    "toml>=0.10.2",
    "loguru>=0.7.2",
    "nbclient>=0.10.2",
    "pandas>=2.2.3",
    "jupytext>=1.17.2",
    "isort>=6.0.1",
    "numpy>=2.3.0",
    "altair>=5.5.0",
    "hdbscan>=0.8.40",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
    "sentence-transformers>=4.1.0",
    "python-dotenv>=1.1.0",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.15",
    "ftfy>=6.3.1",
]

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
